requirements.txt
alabaster==0.7.12
azure-storage-blob==1.5.0
black==19.3b0
coverage==4.5.3
flake8==3.7.7
imapclient==2.1.0
ipython==7.5.0
isort==4.3.18
jupyter==1.0.0
jupyterlab==0.35.6
kaggle==1.5.3
mypy==0.701
pycodestyle==2.5.0
pydocstyle==3.0.0
pylint==2.3.1
pytest==4.4.1
pytest-cov==2.7.1
pyyaml>=4.2b1
Sphinx==2.0.1
watermark==1.8.1

pandas
numpy
nltk
sklearn

## client/updatechecker
schedule==0.6.0